+++ oneflow/python/ops/math_ops.py
@@ -2121,47 +2121,3 @@ def mish(x: oneflow_api.BlobDesc, name: Optional[str] = None,) -> oneflow_api.Bl
     return x * flow.math.tanh(
         flow.math.softplus(x, name=name + "softplus"), name=name + "tanh"
     )
-
-
-@oneflow_export("math.swish")
-def swish(
-    x: oneflow_api.BlobDesc, beta: float = 1.0, name: Optional[str] = None,
-) -> oneflow_api.BlobDesc:
-    r"""The Swish activation function. 
-
-    The equation is: 
-
-    .. math:: 
-
-        out = x * sigmoid(\beta*x)
-
-    For example: 
-
-    .. code-block:: python 
-
-        import oneflow as flow 
-        import oneflow.typing as tp 
-        import numpy as np 
-
-
-        @flow.global_function()
-        def swish_job(x: tp.Numpy.Placeholder(shape=(5, )))->tp.Numpy: 
-            return flow.math.swish(x)
-        x = np.array([-0.5, 0, 0.5, 1, 1.5]).astype(np.float32)
-
-
-        out = swish_job(x)
-        # output [-0.18877034  0.          0.31122968  0.7310586   1.2263618 ]
-    
-    Args:
-        x (oneflow_api.BlobDesc): The input Blob. 
-        beta (float, optional): The smooth factor. Defaults to 1.0.
-        name (Optional[str], optional): The name for the operation. Defaults to None.
-    
-    Returns:
-        oneflow_api.BlobDesc: The result Blob. 
-    """
-    if name is None:
-        name = id_util.UniqueStr("Swish_")
-
-    return x * flow.math.sigmoid(beta * x, name=name + "_sigmoid")
